Converge all commands into one
LP: #1567058
