--- snapcraft/commands/logout.py
@@ -1,47 +1,0 @@
-# -*- Mode:Python; indent-tabs-mode:nil; tab-width:4 -*-
-#
-# Copyright (C) 2016 Canonical Ltd
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License version 3 as
-# published by the Free Software Foundation.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-"""
-snapcraft logout
-
-Forget credentials for Ubuntu One SSO.
-
-Usage:
-  logout [options]
-
-Options:
-  -h --help             show this help message and exit.
-
-"""
-
-import logging
-
-from docopt import docopt
-
-from snapcraft.config import clear_config
-
-
-logger = logging.getLogger(__name__)
-
-
-def main(argv=None):
-    """Forget credentials for Ubuntu One SSO."""
-    argv = argv if argv else []
-    docopt(__doc__, argv=argv)
-
-    logger.info('Clearing credentials for Ubuntu One SSO.')
-    clear_config()
-    logger.info('Credentials cleared.')
